+++ src/mastodon_api/accounts/views.rs
@@ -20,7 +20,6 @@ use mitra_config::{
 use mitra_models::{
     database::{get_database_client, DatabaseError, DbPool},
     posts::queries::get_posts_by_author,
-    profiles::helpers::find_verified_aliases,
     profiles::queries::{
         get_profile_by_acct,
         get_profile_by_id,
@@ -937,30 +936,8 @@ async fn get_account_subscribers(
     Ok(HttpResponse::Ok().json(subscriptions))
 }
 
-#[get("/{account_id}/aliases")]
-async fn get_account_aliases(
-    connection_info: ConnectionInfo,
-    config: web::Data<Config>,
-    db_pool: web::Data<DbPool>,
-    account_id: web::Path<Uuid>,
-) -> Result<HttpResponse, MastodonError> {
-    let db_client = &**get_database_client(&db_pool).await?;
-    let profile = get_profile_by_id(db_client, &account_id).await?;
-    let aliases = find_verified_aliases(db_client, &profile).await?;
-    let base_url = get_request_base_url(connection_info);
-    let instance_url = config.instance_url();
-    let accounts: Vec<Account> = aliases.into_iter()
-        .map(|profile| Account::from_profile(
-            &base_url,
-            &instance_url,
-            profile,
-        ))
-        .collect();
-    Ok(HttpResponse::Ok().json(accounts))
-}
-
 #[get("/{account_id}/aliases/all")]
-async fn get_all_account_aliases(
+async fn get_account_aliases(
     connection_info: ConnectionInfo,
     config: web::Data<Config>,
     db_pool: web::Data<DbPool>,
@@ -1004,5 +981,4 @@ pub fn account_api_scope() -> Scope {
         .service(get_account_following)
         .service(get_account_subscribers)
         .service(get_account_aliases)
-        .service(get_all_account_aliases)
 }